+++ test/web.dart
@@ -5,13 +5,21 @@
 library dartpad.web_test;
 
 import 'package:unittest/html_config.dart';
+import 'package:unittest/unittest.dart';
 
+import 'core/dependencies_test.dart' as dependencies_test;
+import 'core/event_bus_test.dart' as event_bus_test;
 import 'core/keys_test.dart' as keys_test;
+import 'services/common_test.dart' as common_test;
 
 void main() {
   // Set up the test environment.
-  useHtmlConfiguration();
+  // TODO: Use LoggingHtmlConfiguration when available.
+  unittestConfiguration = new HtmlConfiguration(false);
 
   // Define the tests.
+  dependencies_test.defineTests();
+  common_test.defineTests();
+  event_bus_test.defineTests();
   keys_test.defineTests();
 }
