+++ test/web.html
@@ -12,7 +12,6 @@
   </head>
 
   <body>
-    <script src="packages/unittest/test_controller.js"></script>
     <script type="application/dart" src="web.dart"></script>
     <script src="packages/browser/dart.js"></script>
   </body>
